% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAnalysis.R
\name{checkFit}
\alias{checkFit}
\title{Check goodness of fit of a trend model.}
\usage{
checkFit(trend, residuals = TRUE, ...)
}
\arguments{
\item{trend}{A fitted object of class trend.}

\item{residuals}{Should residuals be plotted?}

\item{...}{Further arguments passed to \code{\link[mgcv]{plot.gam}}.}
}
\description{
Produces various goodness of fit plots and diagnostic measures.
}
\details{
The function simply calls \link[mgcv]{plot.gam} and \link[mgcv]{gam.check} on the 
underlying gam model for checking goodness of fit.
}
\author{
Jonas Knape
}
\seealso{
\code{\link[mgcv]{plot.gam}}, \code{\link[mgcv]{gam.check}}
}

