% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAnalysis.R
\name{plot.trend}
\alias{plot.trend}
\title{Plot an estimated trend.}
\usage{
\method{plot}{trend}(
  x,
  baseline = NULL,
  ylab = "abundance index",
  trendCol = "#0072B2",
  lineCol = adjustcolor("#0072B2", alpha.f = 0.05),
  shadeCol = adjustcolor("#0072B2", alpha.f = 0.4),
  incCol = "#009E73",
  decCol = "#d90211",
  plotGrid = TRUE,
  gridCol = "gray80",
  plotLines = FALSE,
  ranefCI = TRUE,
  lwd = 3,
  ...
)
}
\arguments{
\item{x}{A fitted object of class trend.}

\item{baseline}{A single time point, or vector of time points, that are used to define the reference 
level for the index. Can only contain time points for which there are observations in the data.
If baseline is a vector of length larger than one, the estimated mean over all the time 
points is taken as the reference. Defaults to NULL in which case the current baseline of the
x is used (often the first time point in the data).}

\item{ylab}{The label of the y-axis.}

\item{trendCol}{The color of the trend line.}

\item{lineCol}{The color of bootstrapped trend lines, if plotted.}

\item{shadeCol}{The color of the confidence region.}

\item{incCol}{The color of regions where the first or second derivative is significantly increasing.}

\item{decCol}{The color of regions where the first or second derivative is significantly decreasing.}

\item{plotGrid}{If true, grid lines are plotted.}

\item{gridCol}{Color of grid lines.}

\item{plotLines}{If true, the bootstrapped trends are plotted.}

\item{ranefCI}{If true, confidence intervals for random effects are plotted.}

\item{lwd}{Width of the trend line for smooth or loglinear trends.}

\item{...}{Further arguments passed to \code{\link[graphics]{plot.default}}.}

\item{alpha}{The alpha level of confidence intervals.}
}
\description{
The function plots an estimated trend or index, as well as estimates of any temporal random effects included in the
trend term.
}
\details{
Trends and indexes are relative measures and therefore are compared against some reference value.
By default, the first observed time point is used as the reference value.

If the estimated trend contains bootstrap samples, confidence intervals are plotted as well.
For smooth trend models, time periods where the trend is significantly declining or increasing are marked with
a different color (set by arguments \var{decCol} and \var{incCol}). Periods where the second derivative is 
significantly positive or negative are marked by coloured boxes at the bottom of the plot. 


There is an additional option of plotting each of the bootstrapped trends.
}
\author{
Jonas Knape
}
