% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/poptrend-package.r
\docType{package}
\name{poptrend}
\alias{poptrend}
\alias{poptrend-package}
\title{Analyze population trends obtained from survey count data.}
\description{
Provides functions for fitting and analysing trend models of data obtained
from point count surveys.
}
\details{
The package provide functions for estimating smooth and
The package is intended as a simple interface to trend estimation, allowing estimation of
trends accounting for effects of covariates in the form of both smooth terms and random effects.
The model fitting engine is the function \code{\link[mgcv]{gam}} of package \link[mgcv]{mgcv}.
}

