% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAnalysis.R
\name{print.trend}
\alias{print.trend}
\title{Print a trend object.}
\usage{
\method{print}{trend}(x, ...)
}
\arguments{
\item{x}{A trend object.}

\item{...}{Not used.}
}
\description{
Prints basic information about a trend object.
}
\details{
Prints the family, formula and type of trend.
}
\author{
Jonas Knape
}

