% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trendAnalysis.R
\name{simTrend}
\alias{simTrend}
\title{Simulate population survey data.}
\usage{
simTrend(nyear = 30, nsite = 40, mu = 3, timeSD = 0.1)
}
\arguments{
\item{nyear}{The number of years in the simulated survey.}

\item{nsite}{The number of sites in the simulated survey}

\item{mu}{The expected mean of the counts at the start of the survey.}

\item{timeSD}{Standard deviation of annual mean deviation from the trend.}
}
\value{
A data frame containing simulated data.
}
\description{
Simulates count survey data with a non-linear trend, and site and temporal random effects.
The logistic function is used to create a trend the reduces the population size to half over
the time period . The mean of the counts at the start of the
}
\author{
Jonas Knape
}

