% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAnalysis.R
\name{summary.trend}
\alias{summary.trend}
\title{Summary of trend estimates}
\usage{
\method{summary}{trend}(object, baseline = NULL, ...)
}
\arguments{
\item{object}{A trend object returned by \code{\link{ptrend}}.}

\item{baseline}{}

\item{...}{Not used.}
}
\description{
Computes a index estimate for each time point in the survey.
}
\details{
For a smooth or loglinear trend model the function computes an index from
the estimated trend line at each time point in the survey. By default, the reference
value is the first time point. Note that if the trend model was fitted with random 
effects, the random effects are not included in the index. Thus the index refers
to the long-term component of the trend.

For an index trend model the index at each time point is computed.
}
\author{
Jonas Knape
}
