% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendFit.R
\name{updateBaseline}
\alias{updateBaseline}
\title{Update the baseline of a trend object.}
\usage{
updateBaseline(object, baseline = NA, level = 0.95)
}
\arguments{
\item{object}{An object of class trend.}

\item{baseline}{A single time point, or vector of time points, that are used to define the reference 
level for the index. Can only contain time points for which there are observations in the data.
If baseline is a vector of length larger than one, the estimated mean over all the time 
points is taken as the reference. By default, the first time point in the data is taken as the reference.}

\item{level}{The confidence level for uncertainty estimates.}
}
\value{
A trend object with updated baseline and confidence level.
}
\description{
This function can be used to update the baseline for indices computed from a trend object.
}
\author{
Jonas Knape
}
